+++ tests/e2e/cases/generated_c_dce_const/in.ts
@@ -583,6 +583,22 @@ const unused_promise_resolve_adopt_reject_catch_call = Promise.resolve(Promise.r
     .catch(() => "dead_promise_resolve_adopt_reject_catch_callback");
 const unused_promise_resolve_adopt_all_finally_call = Promise.resolve(Promise.all([] as Promise<string>[]))
     .finally(() => "dead_promise_resolve_adopt_all_finally_callback");
+const unused_promise_all_fulfilled_then_call = Promise.all([Promise.resolve("dead_promise_all_fulfilled_source")])
+    .then(() => "dead_promise_all_fulfilled_then_callback");
+const unused_promise_all_rejected_catch_call = Promise.all([Promise.reject<string>("dead_promise_all_rejected_source")])
+    .catch(() => "dead_promise_all_rejected_catch_callback");
+const unused_promise_any_fulfilled_then_call = Promise.any([
+    Promise.reject<string>("dead_promise_any_fulfilled_rejected_source"),
+    Promise.resolve("dead_promise_any_fulfilled_source"),
+])
+    .then(() => "dead_promise_any_fulfilled_then_callback");
+const unused_promise_all_settled_nonempty_then_call = Promise.allSettled([
+    Promise.resolve("dead_promise_all_settled_nonempty_resolve_source"),
+    Promise.reject<string>("dead_promise_all_settled_nonempty_reject_source"),
+])
+    .then(() => "dead_promise_all_settled_nonempty_then_callback");
+const unused_promise_race_fulfilled_then_call = Promise.race([Promise.resolve("dead_promise_race_fulfilled_source")])
+    .then(() => "dead_promise_race_fulfilled_then_callback");
 const unused_new_promise_empty_executor_call = new Promise<string>(() => {
 });
 const unused_new_promise_resolve_call = new Promise<string>((resolve) => resolve("dead_new_promise_resolve"));
@@ -1639,6 +1655,22 @@ Promise.resolve(Promise.reject("top_level_dead_promise_resolve_adopt_reject_sour
     .catch(() => "top_level_dead_promise_resolve_adopt_reject_catch_callback");
 Promise.resolve(Promise.all([] as Promise<string>[]))
     .finally(() => "top_level_dead_promise_resolve_adopt_all_finally_callback");
+Promise.all([Promise.resolve("top_level_dead_promise_all_fulfilled_source")])
+    .then(() => "top_level_dead_promise_all_fulfilled_then_callback");
+Promise.all([Promise.reject<string>("top_level_dead_promise_all_rejected_source")])
+    .catch(() => "top_level_dead_promise_all_rejected_catch_callback");
+Promise.any([
+    Promise.reject<string>("top_level_dead_promise_any_fulfilled_rejected_source"),
+    Promise.resolve("top_level_dead_promise_any_fulfilled_source"),
+])
+    .then(() => "top_level_dead_promise_any_fulfilled_then_callback");
+Promise.allSettled([
+    Promise.resolve("top_level_dead_promise_all_settled_nonempty_resolve_source"),
+    Promise.reject<string>("top_level_dead_promise_all_settled_nonempty_reject_source"),
+])
+    .then(() => "top_level_dead_promise_all_settled_nonempty_then_callback");
+Promise.race([Promise.resolve("top_level_dead_promise_race_fulfilled_source")])
+    .then(() => "top_level_dead_promise_race_fulfilled_then_callback");
 new Promise<string>(() => {
 });
 new Promise<string>((resolve) => resolve("top_level_dead_new_promise_resolve"));
@@ -2672,6 +2704,22 @@ function usedLocal(value: number): number {
         .catch(() => "local_dead_promise_resolve_adopt_reject_catch_callback");
     Promise.resolve(Promise.all([] as Promise<string>[]))
         .finally(() => "local_dead_promise_resolve_adopt_all_finally_callback");
+    Promise.all([Promise.resolve("local_dead_promise_all_fulfilled_source")])
+        .then(() => "local_dead_promise_all_fulfilled_then_callback");
+    Promise.all([Promise.reject<string>("local_dead_promise_all_rejected_source")])
+        .catch(() => "local_dead_promise_all_rejected_catch_callback");
+    Promise.any([
+        Promise.reject<string>("local_dead_promise_any_fulfilled_rejected_source"),
+        Promise.resolve("local_dead_promise_any_fulfilled_source"),
+    ])
+        .then(() => "local_dead_promise_any_fulfilled_then_callback");
+    Promise.allSettled([
+        Promise.resolve("local_dead_promise_all_settled_nonempty_resolve_source"),
+        Promise.reject<string>("local_dead_promise_all_settled_nonempty_reject_source"),
+    ])
+        .then(() => "local_dead_promise_all_settled_nonempty_then_callback");
+    Promise.race([Promise.resolve("local_dead_promise_race_fulfilled_source")])
+        .then(() => "local_dead_promise_race_fulfilled_then_callback");
     new Promise<string>(() => {
     });
     new Promise<string>((resolve) => resolve("local_dead_new_promise_resolve"));